+++ lib/sift/filter.rb
