+++ lib/sift/filter_validator.rb
