+++ lib/sift/filtrator.rb
