+++ lib/sift/parameter.rb
