+++ lib/sift/scope_handler.rb
