+++ lib/sift/subset_comparator.rb
