+++ lib/sift/value_parser.rb
