+++ lib/sift/where_handler.rb
